resolution & fullscreen options in settings
added a dropdown for the resolution and a checkbox for the fullscreen options.

created a settings menu script in order for these options to work. attached script as a component to the canvas in order to have an object reference to be able to use the methods as defined in the script.

not entirely sure if the options work 100% as they should