added sound option in settings
added the option to change the volume using a slider in the settings.

created an audio mixer that will control the master volume. exposed its parameters so the slider in the settings menu is able to control the mixer.

updated the settings menu script to add sound functionality.

seems to work when using Debug.Log to display the volume variable changing but will need to test with actual in-game sound to be 100% sure